+++ test/libsolidity/smtCheckerTests/complex/warn_on_struct.sol
@@ -9,5 +9,6 @@ contract C {
 // ----
 // Warning: (133-143): Unused local variable.
 // Warning: (133-143): Assertion checker does not yet support the type of this variable.
+// Warning: (146-147): Assertion checker does not yet implement type type(struct C.A storage pointer)
 // Warning: (146-163): Assertion checker does not yet implement type struct C.A memory
 // Warning: (146-163): Assertion checker does not yet implement this expression.