+++ Changelog.md
@@ -42,6 +42,7 @@ Compiler Features:
 
 
 Bugfixes:
+ * SMTChecker: Fix internal error when using ``abi.decode``.